Verify the password and set cookie with hashed password.
Let's try to hash(2) the plaintext and an already-hashed password.
This should permit us to keep the password on disk hashed(1).
Not sure if a good idea. An attacker who captured the hashed(1)
password off the disk can use it to create login cookies.
He can't use our login form, but that is hardly helpful.