Switch to unicode internal strings for Fedora 15.
F15 ships with systemd, which, unlike Upstart, does not set LANG.
Apache runs with no LANG, and so do WSGI modules. Because of that,
there is a ton of UnicodeEncodeError exceptions because "'ascii'
codec can't encode characters". All instances of sort blow up,
and all attempts to open or create a file. The latter is most
unfortunate, because it happens even if the filename is manually
encoded and passed as str type. The only solution we found was
to use base64 encoded filenames.