Protect from fault-in-fault with garbled URLs.
Although we try to plug all inputs, sometimes garbage ends in an
argument for an exception. When that happens, str(e) returns a
binary string... which would be fine, except that it blows up
immediately at '%' operator (with a charater ordinary out of range
for 'ascii' codec). Even if we manage to stuff this through by
use of '+', the framework will blow up anyway when it introspects
the returned lists before sending them to network.