Add the "Preload" button
When I started using Slasti from a tablet, it turned out that copying
of random text for the link title was impossible. So, to provide a
small assistance, we add a "Preload" button to the new mark form.
Hitting this button pre-loads the title, using the URL field.

Unfortunately, we found no way to get client to load a random URL,
due to XSS protections. So, we use a server-side indirection,
carefuly. It may yet turn out to be problematic, so suggestions
are welcome.

This patch adds a JavaScript component, in a separate file.
Installation instructions adjusted accordingly.