Fix POST input handling
Some WSGI implementations hang when readline() is called on wsgi.input
and there's no linebreak in the input. Try to avoid this situation by
reading exactly CONTENT_LENGTH bytes if CONTENT_LENGTH is available.
If it's missing, fall back to the readline() implementation.

This behavior likely depends on the WSGI server: Apache seems to signal
EOF as needed. In contrast, cherrypy's WSGI server just hung on
readline().