Read marks as unicode
This is the meat of Christian's all-unicode patch. It comes after
the templates, but it must be pulled forward, because it is impossible
to tinker with the templates using the available mark corpii, which
contain international tags and titles.

Clearly it's a bug: we write marks using UTF-8 writer, so we must
read them using UTF-8 reader. The current code miraclously works by
using safestr(), but the dictionaries of templates mix everything up,
thus making the problem apparent.