Splice TemplateNodeElem into the templating
This is a shim that comes about for the rework of looping.
Christian rolled everything into a string first, and only then
reconstructed #if and #for. We worked around #if without changing
much, but that is not enough for loops. Solution is to keep the
template structure until the expansion time, so that sub-templates
could be expanded in a loop.

For now, the skeleton is in the code, but not enabled yet.